src/requirements.txt
python-telegram-bot==20.8
boto3==1.34.41
pytest==8.1.1